OPE-49 convert agent page to sidebar sub-pages pattern
- AgentLayout with Outlet for nested routes (mirrors SettingsLayout)
- AgentBasicSettingsPage as default sub-page (Greeting + Default Language)
- Sidebar nav: Agent now collapsible with 'Basic Settings' sub-item
- Removed old AgentPage.tsx and accordion.tsx
