--- apps/web/src/components/ui/accordion.tsx
@@ -1,91 +1,0 @@
-import * as React from "react";
-import { Accordion as AccordionPrimitive } from "@base-ui/react/accordion";
-import { ChevronDown } from "lucide-react";
-
-import { cn } from "@/lib/utils";
-
-function Accordion({
-  className,
-  ...props
-}: AccordionPrimitive.Root.Props & { className?: string }) {
-  return (
-    <AccordionPrimitive.Root
-      data-slot="accordion"
-      className={cn("space-y-3", className)}
-      {...props}
-    />
-  );
-}
-
-function AccordionItem({
-  className,
-  ...props
-}: AccordionPrimitive.Item.Props & { className?: string }) {
-  return (
-    <AccordionPrimitive.Item
-      data-slot="accordion-item"
-      className={cn(
-        "rounded-xl border border-border/70 bg-card/90 shadow-sm",
-        className,
-      )}
-      {...props}
-    />
-  );
-}
-
-function AccordionHeader({
-  ...props
-}: AccordionPrimitive.Header.Props) {
-  return (
-    <AccordionPrimitive.Header
-      data-slot="accordion-header"
-      {...props}
-    />
-  );
-}
-
-function AccordionTrigger({
-  className,
-  children,
-  ...props
-}: AccordionPrimitive.Trigger.Props & { className?: string }) {
-  return (
-    <AccordionPrimitive.Trigger
-      data-slot="accordion-trigger"
-      className={cn(
-        "flex w-full items-center justify-between gap-4 px-6 py-4 text-left text-sm font-semibold transition-colors outline-none hover:bg-muted/40 focus-visible:ring-2 focus-visible:ring-ring/50 disabled:pointer-events-none disabled:opacity-50 [&[data-panel-open]>svg]:rotate-180",
-        className,
-      )}
-      {...props}
-    >
-      {children}
-      <ChevronDown className="size-4 shrink-0 text-muted-foreground transition-transform duration-200" />
-    </AccordionPrimitive.Trigger>
-  );
-}
-
-function AccordionPanel({
-  className,
-  children,
-  ...props
-}: AccordionPrimitive.Panel.Props & { className?: string }) {
-  return (
-    <AccordionPrimitive.Panel
-      data-slot="accordion-panel"
-      className={cn("overflow-hidden", className)}
-      {...props}
-    >
-      <div className={cn("border-t border-border/50 px-6 pb-6 pt-4", className)}>
-        {children}
-      </div>
-    </AccordionPrimitive.Panel>
-  );
-}
-
-export {
-  Accordion,
-  AccordionItem,
-  AccordionHeader,
-  AccordionTrigger,
-  AccordionPanel,
-};
