+++ apps/web/src/features/agent/AgentBasicSettingsPage.tsx
@@ -1,0 +1,110 @@
+import { type FormEvent, useEffect, useState } from "react";
+import { useMutation, useQuery } from "convex/react";
+import type { RuntimeLocale } from "@ai-receptionist/shared";
+import { useTranslation } from "react-i18next";
+
+import type { Id } from "../../../../../convex/_generated/dataModel";
+import { api } from "../../../../../convex/_generated/api";
+import { Button } from "@/components/ui/button";
+import { Input } from "@/components/ui/input";
+import { Label } from "@/components/ui/label";
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from "@/components/ui/select";
+
+type AgentBasicSettingsPageProps = {
+  businessId: Id<"businesses">;
+};
+
+export function AgentBasicSettingsPage({ businessId }: AgentBasicSettingsPageProps) {
+  const { t } = useTranslation(["agent", "common"]);
+  const configuration = useQuery(api.businesses.catalog.getBusinessConfiguration, {
+    businessId,
+  });
+  const saveProfile = useMutation(api.ai.context.snapshots.updateReceptionistProfile);
+
+  const [greeting, setGreeting] = useState("");
+  const [defaultLocale, setDefaultLocale] = useState<RuntimeLocale>("en");
+  const [status, setStatus] = useState<string | null>(null);
+  const [isSaving, setIsSaving] = useState(false);
+
+  useEffect(() => {
+    const profile = configuration?.profile;
+    if (!profile) {
+      return;
+    }
+    setGreeting(profile.greeting);
+    setDefaultLocale(configuration.business?.defaultLocale ?? "en");
+  }, [configuration]);
+
+  async function handleSubmit(event: FormEvent<HTMLFormElement>): Promise<void> {
+    event.preventDefault();
+    setIsSaving(true);
+    setStatus(null);
+    try {
+      await saveProfile({
+        businessId,
+        defaultLocale,
+        greeting,
+        tone: configuration?.profile?.tone ?? "",
+        summary: configuration?.profile?.summary ?? "",
+        bookingPolicy: configuration?.profile?.bookingPolicy ?? "",
+        transferMode: configuration?.profile?.transferMode ?? "on_request",
+      });
+      setStatus(t("agent:actions.saved"));
+    } finally {
+      setIsSaving(false);
+    }
+  }
+
+  return (
+    <div className="w-full max-w-xl">
+      <form className="space-y-6" onSubmit={(event) => void handleSubmit(event)}>
+        <div className="space-y-4">
+          <div className="space-y-2">
+            <Label htmlFor="agent-greeting">{t("agent:fields.greeting.label")}</Label>
+            <Input
+              id="agent-greeting"
+              placeholder={t("agent:fields.greeting.placeholder")}
+              value={greeting}
+              onChange={(event) => setGreeting(event.target.value)}
+            />
+            <p className="text-xs text-muted-foreground">
+              {t("agent:fields.greeting.hint")}
+            </p>
+          </div>
+
+          <div className="space-y-2">
+            <Label htmlFor="agent-default-language">{t("agent:fields.defaultLanguage.label")}</Label>
+            <Select
+              value={defaultLocale}
+              onValueChange={(value) => setDefaultLocale((value as RuntimeLocale | "") || "en")}
+            >
+              <SelectTrigger id="agent-default-language">
+                <SelectValue />
+              </SelectTrigger>
+              <SelectContent>
+                <SelectItem value="en">{t("common:language.english")}</SelectItem>
+                <SelectItem value="fr">{t("common:language.french")}</SelectItem>
+              </SelectContent>
+            </Select>
+            <p className="text-xs text-muted-foreground">
+              {t("agent:fields.defaultLanguage.hint")}
+            </p>
+          </div>
+        </div>
+
+        <div className="flex items-center gap-3">
+          <Button disabled={isSaving} type="submit">
+            {isSaving ? t("agent:actions.saving") : t("agent:actions.save")}
+          </Button>
+          {status ? <span className="text-sm text-muted-foreground">{status}</span> : null}
+        </div>
+      </form>
+    </div>
+  );
+}
